refactor: split Computation into layered modules
computeengine.py was 2433 lines defining a Computation class with 126
methods. It was the only module in src/loman/ that radon mi ranked below
B, scoring C (0.00) — the floor of the scale.

The class is now assembled from a chain of layers, each in its own module
and each calling only the layers beneath it:

  base          shared state every layer operates on
  subscription  batching and dispatch of change notifications
  state         node state transitions and the state index
  query         read-only accessors over the graph
  attributes    tags, styles and metadata
  execution     scheduling and running calculations
  persistence   copying, and file round-trips
  mutation      adding, removing, rewiring and populating nodes

A linear chain rather than sibling mixins because the dependencies turned
out to be acyclic: it makes the direction explicit, keeps the import graph
linear, and lets the type checker resolve every cross-layer call.

subscription sits directly above base because the lowest layer that has to
report a change is state — every transition marks a node, and the public
mutations further up decide where a batch begins and ends. The
_notifies_subscribers decorator therefore has to be importable by state and
by everything above it, and nothing below it needs to know notification
exists.

Also adds values.py, a leaf module holding the types the engine and the
serialization layer both need (Error, NodeData, TimingData, ConstantValue,
_ParameterType, _ParameterItem). serialization/computation.py now imports
Error and _ParameterType from there at module scope instead of reaching
into computeengine at call time, removing two of the three deferred
imports behind the engine<->serialization cycle. The third — constructing
Computation in _from_dict — is untouched and remains for #344.

Behaviour is unchanged:

- Public API verified identical against master by diffing dir(loman),
  loman.__all__ and every Computation member signature. The only
  differences are annotation reprs following types to their new module,
  and copy()/read_dill() now returning Self. The decorated mutations keep
  their signatures because _notifies_subscribers wraps with functools.wraps.
- Types whose instances live inside the DAG are re-exported from
  computeengine under their historical names, so dill files written by
  earlier versions still load. Verified by writing with pre-split code
  and reading with the split code. ComputationEvent is re-exported for the
  same reason loman.ui imports it from computeengine.
- write_dill_old detaches __getstate__/__setstate__ from PersistenceMixin,
  which now defines them, rather than from Computation, which inherits.
- Subscription state is declared on SubscriptionMixin and populated by
  Computation.__init__, the same convention ComputationBase already uses.
  It is neither copied nor serialized: copy() builds a fresh Computation,
  so a copy still starts with no subscribers.
- The logger name stays "loman.computeengine" in the moved code.

radon mi for computeengine.py: C (0.00) -> A (46.50). Every module is now
A except visualization.py, unchanged at B (17.48).

Gates: fmt, typecheck, docs-coverage (100%), deptry, rhiza-test and test
(984 passed, 100% coverage on 3442 statements) all pass. The benchmark and
stress suites the test gate excludes pass too, for 1012 in total.

Closes #345
