+++ src/loman/computeengine.py
@@ -1,218 +1,87 @@
-"""Core computation engine for dependency-aware calculation graphs."""
+"""Core computation engine for dependency-aware calculation graphs.
 
-import contextlib
+:class:`Computation` is the top of a stack of layers, each in its own module
+and each calling only the layers beneath it — lowest first:
+
+=========================  =====================================================
+:mod:`loman.base`          the shared state all layers operate on
+:mod:`loman.subscription`  batching and dispatch of change notifications
+:mod:`loman.state`         node state transitions and the state index
+:mod:`loman.query`         read-only accessors over the graph
+:mod:`loman.attributes`    tags, styles and metadata
+:mod:`loman.execution`     scheduling and running calculations
+:mod:`loman.persistence`   copying, and file round-trips
+:mod:`loman.mutation`      adding, removing, rewiring and populating nodes
+=========================  =====================================================
+
+What is left in this module is the object's construction, the class-definition
+DSL that builds computations out of decorated classes, and the delegation to
+planning and visualization.
+"""
+
 import functools
 import inspect
 import logging
-import traceback
-import types
-import warnings
-import weakref
 from collections import defaultdict
-from collections.abc import Callable, Iterable, Mapping
-from concurrent.futures import FIRST_COMPLETED, Executor, ThreadPoolExecutor, wait
+from collections.abc import Callable, Iterable
+from concurrent.futures import Executor, ThreadPoolExecutor
 from dataclasses import dataclass, field
-from datetime import UTC, datetime
-from enum import Enum
-from types import MappingProxyType
-from typing import TYPE_CHECKING, Any, BinaryIO, TextIO, TypeVar, cast, overload
+from typing import TYPE_CHECKING, Any, TypeVar, overload
 
 if TYPE_CHECKING:
-    from .serialization.computation import ComputationSerializer
     from .ui import ComputationWidget
 
 import decorator
-import dill  # nosec B403
 import networkx as nx
-import pandas as pd
 
 from .compat import get_signature
-from .consts import EdgeAttributes, NodeAttributes, NodeTransformations, States, SystemTags
-from .exception import (
-    CannotInsertToPlaceholderNodeException,
-    ComputationError,
-    LoopDetectedException,
-    MapException,
-    NodeAlreadyExistsException,
-    NonExistentNodeException,
-    ValidationError,
-)
-from .graph_utils import topological_sort
-from .nodekey import Name, Names, NodeKey, names_to_node_keys, node_keys_to_names, to_nodekey
+from .consts import NodeTransformations, States, SystemTags
+from .exception import NonExistentNodeException
+from .mutation import MutationMixin, identity_function
+from .nodekey import Name, Names, NodeKey, names_to_node_keys, to_nodekey
 from .planning import ExecutionPlan, ValidationReport, create_execution_plan, validate_graph
-from .util import AttributeView, apply1, apply_n, as_iterable, value_eq
+from .subscription import ComputationEvent
+from .util import AttributeView
+from .values import C, ConstantValue, Error, NodeData, TimingData
+
+# Re-exported under their historical names. These are not merely a convenience:
+# instances of them are stored inside the DAG (``Error`` as a node value,
+# ``_ParameterType`` inside every edge's PARAM attribute), and pickle records
+# the module path it first saw. Dropping these names would make dill files
+# written by earlier versions unreadable by ``Computation.read_dill``.
+from .values import _ParameterItem as _ParameterItem
+from .values import _ParameterType as _ParameterType
 from .visualization import GraphView, NodeFormatter
 
 LOG = logging.getLogger("loman.computeengine")
 
 F = TypeVar("F", bound=Callable[..., Any])
 
-
-@dataclass
-class Error:
-    """Container for error information during computation."""
-
-    exception: Exception
-    traceback: str
-
-
-@dataclass
-class NodeData:
-    """Data associated with a computation node."""
-
-    state: States
-    value: object
-
-
-@dataclass
-class TimingData:
-    """Timing information for computation execution."""
-
-    start: datetime
-    end: datetime
-    duration: float
-
-
-@dataclass(frozen=True)
-class ComputationEvent:
-    """A batched notification describing a mutation to a computation.
-
-    Subscribers receive one event after each outermost public mutation, even
-    when that operation performs many internal state transitions. Values are
-    deliberately excluded: consumers can fetch a changed value lazily from
-    :attr:`computation` without copying large objects into every event.
-
-    :ivar computation: The live computation that produced the event. It is not a
-        snapshot, and continues to change after the event is delivered.
-    :ivar revision: Monotonic counter, matching :attr:`Computation.revision` at
-        the moment the event was published.
-    :ivar changed_nodes: Nodes whose state changed during the operation. When
-        :attr:`graph_changed` is true this is *not* a complete description of the
-        change, because adding, deleting or renaming nodes and altering tags or
-        styles need not change any node's state. Consumers reacting to a
-        structural event should re-read the graph rather than trusting this set.
-    :ivar states: The state of each entry in :attr:`changed_nodes` that still
-        exists, as of publication. Deleted nodes are absent.
-    :ivar graph_changed: True when the structure or presentation of the graph
-        changed, so any cached rendering of it is stale.
-    """
-
-    computation: "Computation"
-    revision: int
-    changed_nodes: frozenset[NodeKey]
-    states: Mapping[NodeKey, States]
-    graph_changed: bool = False
-
-
-ComputationSubscriber = Callable[[ComputationEvent], None]
-
-#: Cap on how many times subscriber-initiated mutations may cascade within one
-#: dispatch before Loman gives up. A well-behaved subscriber settles in one or
-#: two rounds; anything beyond this is a feedback loop rather than useful work.
-_MAX_NOTIFICATION_CASCADES = 16
-
-
-class _Subscription:
-    """One registered subscriber, held weakly when that is safe to do.
-
-    A callback bound to an object is held weakly, so subscribing a widget's
-    handler does not keep the widget alive for the lifetime of the computation.
-    Everything with no object behind it --- plain functions, lambdas, callable
-    objects, :func:`functools.partial` --- is held strongly, because callers
-    routinely pass a closure they retain no other reference to and holding
-    those weakly would collect them immediately.
-
-    "Bound to an object" means carrying a ``__self__``, which covers methods
-    written in Python and those written in C alike. The two need different
-    holders: :class:`weakref.WeakMethod` needs a ``__func__`` to rebind
-    against, and C methods have none, so those are held as a weak reference to
-    the owner plus the attribute name.
-
-    A few owners cannot be weakly referenced at all --- :class:`list`,
-    :class:`dict` and :class:`bytearray` among them --- so ``some_list.append``
-    falls back to a strong reference. That is a limitation of the type rather
-    than a decision here, and it errs towards a subscription that keeps
-    delivering rather than one that silently stops.
-    """
-
-    __slots__ = ("_name", "_owner", "_strong", "_weak")
-
-    def __init__(self, callback: ComputationSubscriber) -> None:
-        """Wrap ``callback``, choosing weak or strong ownership."""
-        self._weak: weakref.WeakMethod | None = None
-        self._owner: weakref.ref[Any] | None = None
-        self._name: str = ""
-        self._strong: ComputationSubscriber | None = None
-        if inspect.ismethod(callback):
-            self._weak = weakref.WeakMethod(callback)
-            return
-        owner = getattr(callback, "__self__", None)
-        name = getattr(callback, "__name__", None)
-        if owner is not None and name:
-            try:
-                self._owner = weakref.ref(owner)
-            except TypeError:
-                # list, dict, bytearray and friends support no weak references.
-                self._strong = callback
-            else:
-                self._name = name
-            return
-        self._strong = callback
-
-    def resolve(self) -> ComputationSubscriber | None:
-        """Return the callback, or ``None`` once a weakly held owner is gone."""
-        if self._strong is not None:
-            return self._strong
-        if self._weak is not None:
-            return self._weak()
-        if self._owner is None:  # pragma: no cover - constructor covers all paths
-            return None
-        owner = self._owner()
-        return None if owner is None else getattr(owner, self._name, None)
-
-
-def _notifies_subscribers(*, graph_changed: bool = False) -> Callable[[F], F]:
-    """Batch changes made by a public mutation and notify on completion.
-
-    With no subscribers attached the wrapper is a straight pass-through, so
-    ordinary use of Loman pays nothing for the notification machinery.
-    """
-
-    def decorate(method: F) -> F:
-        @functools.wraps(method)
-        def wrapped(self: "Computation", *args: Any, **kwargs: Any) -> Any:
-            if self._change_depth == 0 and not self._subscriptions:
-                return method(self, *args, **kwargs)
-            self._change_depth += 1
-            try:
-                result = method(self, *args, **kwargs)
-                if graph_changed:
-                    self._pending_graph_changed = True
-                return result
-            finally:
-                self._change_depth -= 1
-                if self._change_depth == 0:
-                    self._publish_pending_events()
-
-        return cast("F", wrapped)
-
-    return decorate
-
-
-class _ParameterType(Enum):
-    """Internal enum for distinguishing positional and keyword parameters."""
-
-    ARG = 1
-    KWD = 2
-
-
-@dataclass
-class _ParameterItem:
-    """Internal container for parameter information during computation."""
-
-    type: _ParameterType
-    name: int | str
-    value: object
+# Re-exported for backwards compatibility: these types moved to loman.values,
+# loman.subscription and loman.mutation when Computation was split into mixins,
+# but they have always been importable from loman.computeengine.
+__all__ = [
+    "Block",
+    "C",
+    "CalcNode",
+    "Computation",
+    "ComputationEvent",
+    "ConstantValue",
+    "Error",
+    "InputNode",
+    "Node",
+    "NodeData",
+    "NodeKey",
+    "NullObject",
+    "TimingData",
+    "block",
+    "calc_node",
+    "computation_factory",
+    "identity_function",
+    "input_node",
+    "node",
+    "populate_computation_from_class",
+]
 
 
 def _node(func: Callable[..., Any], *args: Any, **kws: Any) -> Any:  # pragma: no cover
@@ -233,16 +102,6 @@ def node(comp: "Computation", name: Name | None = None, *args: Any, **kw: Any) -
         return result
 
     return inner
-
-
-@dataclass()
-class ConstantValue:
-    """Container for constant values in computations."""
-
-    value: object
-
-
-C = ConstantValue
 
 
 class Node:
@@ -381,34 +240,6 @@ def computation_factory(
     return wrap(maybe_cls)
 
 
-def _eval_node(
-    name: NodeKey,
-    f: Callable[..., Any],
-    args: list[Any],
-    kwds: dict[str, Any],
-    raise_exceptions: bool,
-) -> tuple[Any, Exception | None, str | None, datetime, datetime]:
-    """To make multiprocessing work, this function must be standalone so that pickle works."""
-    exc: Exception | None = None
-    tb: str | None = None
-    start_dt = datetime.now(UTC)
-    try:
-        logging.debug("Running " + str(name))
-        value = f(*args, **kwds)
-        logging.debug("Completed " + str(name))
-    except Exception as e:
-        value = None
-        exc = e
-        tb = traceback.format_exc()
-        if raise_exceptions:
-            raise
-    end_dt = datetime.now(UTC)
-    return value, exc, tb, start_dt, end_dt
-
-
-_MISSING_VALUE_SENTINEL = object()
-
-
 class NullObject:
     """Debug helper object that raises exceptions for all attribute/item access."""
 
@@ -454,12 +285,7 @@ class NullObject:
         return "<NullObject>"
 
 
-def identity_function(x: Any) -> Any:
-    """Return the input value unchanged."""
-    return x
-
-
-class Computation:
+class Computation(MutationMixin):
     """A computation graph that manages dependencies and calculations.
 
     The Computation class provides a framework for building and executing
@@ -505,119 +331,12 @@ class Computation:
         self.src = self.get_attribute_view_for_path(NodeKey.root(), self.print_source, self.print_source)
         self._tag_map: defaultdict[str, set[NodeKey]] = defaultdict(set)
         self._state_map: dict[States, set[NodeKey]] = {state: set() for state in States}
-        self._subscriptions: list[_Subscription] = []
+        self._subscriptions = []
         self._revision = 0
         self._change_depth = 0
         self._publishing = False
-        self._pending_changed_nodes: set[NodeKey] = set()
+        self._pending_changed_nodes = set()
         self._pending_graph_changed = False
-
-    @property
-    def revision(self) -> int:
-        """Return the revision number of the most recently published change."""
-        return self._revision
-
-    def subscribe(self, callback: ComputationSubscriber) -> Callable[[], None]:
-        """Subscribe to batched computation changes.
-
-        Subscribers are notified in registration order, synchronously, on the
-        thread that completes the outermost public mutation. A subscriber that
-        raises is logged and skipped; it never interrupts the mutation or the
-        other subscribers. A subscriber that itself mutates the computation
-        causes a further event to be published once the current round finishes,
-        up to a bounded number of cascades.
-
-        A callback with an object behind it --- anything carrying a
-        ``__self__``, whether written in Python or in C --- is held weakly, so
-        subscribing ``obj.handler`` or ``events.append`` does not keep the
-        owner alive; callers must retain it themselves. Everything else ---
-        plain functions, lambdas, callable objects, :func:`functools.partial`
-        --- is held strongly until unsubscribed, because callers commonly pass
-        a throwaway closure that nothing else references.
-
-        The exception is an owner that supports no weak references at all, such
-        as :class:`list`, :class:`dict` and :class:`bytearray`. There
-        ``some_list.append`` falls back to a strong reference, which is a
-        limitation of the type rather than a choice, and errs towards a
-        subscription that keeps delivering over one that silently stops.
-
-        Subscriptions are not copied by :meth:`copy` and are not serialized.
-
-        :param callback: Function accepting a :class:`ComputationEvent`.
-        :return: An idempotent, no-argument unsubscribe function.
-        """
-        if not callable(callback):
-            msg = "callback must be callable"
-            raise TypeError(msg)
-        subscription = _Subscription(callback)
-        self._subscriptions.append(subscription)
-
-        def unsubscribe() -> None:
-            """Remove this subscription, ignoring repeat calls."""
-            with contextlib.suppress(ValueError):
-                self._subscriptions.remove(subscription)
-
-        return unsubscribe
-
-    def _mark_changed(self, *node_keys: NodeKey) -> None:
-        """Record nodes changed by the current public mutation."""
-        if self._subscriptions:
-            self._pending_changed_nodes.update(node_keys)
-
-    def _take_pending_event(self) -> ComputationEvent:
-        """Consume the batched changes and turn them into one event."""
-        changed_nodes = frozenset(self._pending_changed_nodes)
-        graph_changed = self._pending_graph_changed
-        self._pending_changed_nodes.clear()
-        self._pending_graph_changed = False
-        self._revision += 1
-        states = {
-            node_key: self.dag.nodes[node_key][NodeAttributes.STATE]
-            for node_key in changed_nodes
-            if node_key in self.dag
-        }
-        return ComputationEvent(self, self._revision, changed_nodes, MappingProxyType(states), graph_changed)
-
-    def _publish_pending_events(self) -> None:
-        """Publish batched events, including any a subscriber triggers in turn.
-
-        Re-entrant calls return immediately: the dispatch loop already running
-        picks up whatever the subscriber changed, so a subscriber that mutates
-        the computation cannot recurse into the stack.
-        """
-        if self._publishing:
-            return
-        self._publishing = True
-        try:
-            for _ in range(_MAX_NOTIFICATION_CASCADES):
-                if not self._pending_changed_nodes and not self._pending_graph_changed:
-                    return
-                self._dispatch(self._take_pending_event())
-            if self._pending_changed_nodes or self._pending_graph_changed:
-                LOG.error(
-                    "Computation subscribers kept mutating the computation after %s rounds; "
-                    "discarding further notifications to break the loop",
-                    _MAX_NOTIFICATION_CASCADES,
-                )
-                self._pending_changed_nodes.clear()
-                self._pending_graph_changed = False
-        finally:
-            self._publishing = False
-
-    def _dispatch(self, event: ComputationEvent) -> None:
-        """Deliver one event to every live subscriber, isolating failures."""
-        dead = False
-        for subscription in tuple(self._subscriptions):
-            callback = subscription.resolve()
-            if callback is None:
-                dead = True
-                continue
-            try:
-                callback(event)
-            except Exception:
-                LOG.exception("Computation subscriber failed at revision %s", event.revision)
-        if dead:
-            self._subscriptions = [s for s in self._subscriptions if s.resolve() is not None]
 
     def get_attribute_view_for_path(
         self, nodekey: NodeKey, get_one_func: Callable[[Name], Any], get_many_func: Callable[[Name | Names], Any]
@@ -649,617 +368,6 @@ class Computation:
 
         return AttributeView(node_func, get_one_func_for_path, get_many_func_for_path)
 
-    def _get_names_for_state(self, state: States) -> set[Name]:
-        """Get node names that have a specific state."""
-        return set(node_keys_to_names(self._state_map[state]))
-
-    def _get_tags_for_state(self, tag: str) -> set[Name]:
-        """Get node names that have a specific tag."""
-        return set(node_keys_to_names(self._tag_map[tag]))
-
-    def _process_function_args(self, node_key: NodeKey, node: dict[str, Any], args: list[Any] | None) -> int:
-        """Process positional arguments for a function node."""
-        args_count = 0
-        if args:
-            args_count = len(args)
-            for i, arg in enumerate(args):
-                if isinstance(arg, ConstantValue):
-                    node[NodeAttributes.ARGS][i] = arg.value
-                else:
-                    input_vertex_name = arg
-                    input_vertex_node_key = to_nodekey(input_vertex_name)
-                    if not self.dag.has_node(input_vertex_node_key):
-                        self.dag.add_node(input_vertex_node_key, **{NodeAttributes.STATE: States.PLACEHOLDER})
-                        self._state_map[States.PLACEHOLDER].add(input_vertex_node_key)
-                    self.dag.add_edge(
-                        input_vertex_node_key, node_key, **{EdgeAttributes.PARAM: (_ParameterType.ARG, i)}
-                    )
-        return args_count
-
-    def _build_param_map(
-        self,
-        func: Callable[..., Any],
-        node_key: NodeKey,
-        args_count: int,
-        kwds: dict[str, Any] | None,
-        inspect: bool,
-    ) -> tuple[dict[str, Any], list[str]]:
-        """Build parameter map for function node."""
-        param_map: dict[str, Any] = {}
-        default_names: list[str] = []
-
-        if inspect:
-            signature = get_signature(func)
-            if not signature.has_var_args:
-                for param_name in signature.kwd_params[args_count:]:
-                    if kwds is not None and param_name in kwds:
-                        param_source = kwds[param_name]
-                    else:
-                        param_source = node_key.parent.join_parts(param_name)
-                    param_map[param_name] = param_source
-            if signature.has_var_kwds and kwds is not None:
-                for param_name, param_source in kwds.items():
-                    param_map[param_name] = param_source
-            default_names = signature.default_params
-        else:
-            if kwds is not None:
-                for param_name, param_source in kwds.items():
-                    param_map[param_name] = param_source
-
-        return param_map, default_names
-
-    def _process_function_kwds(
-        self, node_key: NodeKey, node: dict[str, Any], param_map: dict[str, Any], default_names: list[str]
-    ) -> None:
-        """Process keyword arguments for a function node."""
-        for param_name, param_source in param_map.items():
-            if isinstance(param_source, ConstantValue):
-                node[NodeAttributes.KWDS][param_name] = param_source.value
-            else:
-                in_node_name = param_source
-                in_node_key = to_nodekey(in_node_name)
-                if not self.dag.has_node(in_node_key):
-                    if param_name in default_names:
-                        continue
-                    else:
-                        self.dag.add_node(in_node_key, **{NodeAttributes.STATE: States.PLACEHOLDER})
-                        self._state_map[States.PLACEHOLDER].add(in_node_key)
-                self.dag.add_edge(in_node_key, node_key, **{EdgeAttributes.PARAM: (_ParameterType.KWD, param_name)})
-
-    @_notifies_subscribers(graph_changed=True)
-    def add_node(
-        self,
-        name: Name,
-        func: Callable[..., Any] | None = None,
-        *,
-        args: list[Any] | None = None,
-        kwds: dict[str, Any] | None = None,
-        value: Any = _MISSING_VALUE_SENTINEL,
-        converter: Callable[[Any], Any] | None = None,
-        serialize: bool = True,
-        inspect: bool = True,
-        group: str | None = None,
-        tags: Iterable[str] | None = None,
-        style: str | None = None,
-        executor: str | None = None,
-        metadata: dict[str, Any] | None = None,
-    ) -> None:
-        """Adds or updates a node in a computation.
-
-        :param name: Name of the node to add. This may be any hashable object.
-        :param func: Function to use to calculate the node if the node is a calculation node. By default, the input
-            nodes to the function will be implied from the names of the function parameters. For example, a
-            parameter called ``a`` would be taken from the node called ``a``. This can be modified with the
-            ``kwds`` parameter.
-        :type func: Function, default None
-        :param args: Specifies a list of nodes that will be used to populate arguments of the function positionally
-            for a calculation node. e.g. If args is ``['a', 'b', 'c']`` then the function would be called with
-            three parameters, taken from the nodes 'a', 'b' and 'c' respectively.
-        :type args: List, default None
-        :param kwds: Specifies a mapping from parameter name to the node that should be used to populate that
-            parameter when calling the function for a calculation node. e.g. If args is ``{'x': 'a', 'y': 'b'}``
-            then the function would be called with parameters named 'x' and 'y', and their values would be taken
-            from nodes 'a' and 'b' respectively. Each entry in the dictionary can be read as "take parameter
-            [key] from node [value]".
-        :type kwds: Dictionary, default None
-        :param value: If given, the value is inserted into the node, and the node state set to UPTODATE.
-        :type value: default None
-        :param serialize: Whether the node should be serialized. Some objects cannot be serialized, in which
-            case, set serialize to False
-        :type serialize: boolean, default True
-        :param inspect: Whether to use introspection to determine the arguments of the function, which can be
-            slow. If this is not set, kwds and args must be set for the function to obtain parameters.
-        :type inspect: boolean, default True
-        :param group: Subgraph to render node in
-        :type group: default None
-        :param tags: Set of tags to apply to node
-        :type tags: Iterable
-        :param styles: Style to apply to node
-        :type styles: String, default None
-        :param executor: Name of executor to run node on
-        :type executor: string
-        """
-        node_key = to_nodekey(name)
-        LOG.debug(f"Adding node {node_key}")
-        has_value = value is not _MISSING_VALUE_SENTINEL
-        if value is _MISSING_VALUE_SENTINEL:
-            value = None
-        if tags is None:
-            tags = []
-
-        self.dag.add_node(node_key)
-        pred_edges = [(p, node_key) for p in self.dag.predecessors(node_key)]
-        self.dag.remove_edges_from(pred_edges)
-        node = self.dag.nodes[node_key]
-
-        if metadata is None:
-            if node_key in self._metadata:
-                del self._metadata[node_key]
-        else:
-            self._metadata[node_key] = metadata
-
-        self._set_state_and_literal_value(node_key, States.UNINITIALIZED, None, require_old_state=False)
-
-        node[NodeAttributes.TAG] = set()
-        node[NodeAttributes.STYLE] = style
-        node[NodeAttributes.GROUP] = group
-        node[NodeAttributes.ARGS] = {}
-        node[NodeAttributes.KWDS] = {}
-        node[NodeAttributes.FUNC] = None
-        node[NodeAttributes.EXECUTOR] = executor
-        node[NodeAttributes.CONVERTER] = converter
-
-        if func:
-            node[NodeAttributes.FUNC] = func
-            args_count = self._process_function_args(node_key, node, args)
-            param_map, default_names = self._build_param_map(func, node_key, args_count, kwds, inspect)
-            self._process_function_kwds(node_key, node, param_map, default_names)
-            self._set_descendents(node_key, States.STALE)
-
-        if has_value:
-            self._set_uptodate(node_key, value)
-        if node[NodeAttributes.STATE] == States.UNINITIALIZED:
-            self._try_set_computable(node_key)
-        self.set_tag(node_key, tags)
-        if serialize:
-            self.set_tag(node_key, SystemTags.SERIALIZE)
-
-    def _refresh_maps(self) -> None:
-        """Refresh internal tag and state maps from node data."""
-        self._tag_map.clear()
-        for state in States:
-            self._state_map[state].clear()
-        for node_key in self._node_keys():
-            state = self.dag.nodes[node_key][NodeAttributes.STATE]
-            self._state_map[state].add(node_key)
-            tags = self.dag.nodes[node_key].get(NodeAttributes.TAG, set())
-            for tag in tags:
-                self._tag_map[tag].add(node_key)
-
-    def _set_tag_one(self, name: Name, tag: str) -> None:
-        """Set a single tag on a single node."""
-        node_key = to_nodekey(name)
-        self.dag.nodes[node_key][NodeAttributes.TAG].add(tag)
-        self._tag_map[tag].add(node_key)
-
-    @_notifies_subscribers(graph_changed=True)
-    def set_tag(self, name: Name | Names, tag: str | Iterable[str]) -> None:
-        """Set tags on a node or nodes. Ignored if tags are already set.
-
-        :param name: Node or nodes to set tag for
-        :param tag: Tag to set
-        """
-        apply_n(self._set_tag_one, name, tag)
-
-    def _clear_tag_one(self, name: Name, tag: str) -> None:
-        """Clear a single tag from a single node."""
-        node_key = to_nodekey(name)
-        self.dag.nodes[node_key][NodeAttributes.TAG].discard(tag)
-        self._tag_map[tag].discard(node_key)
-
-    @_notifies_subscribers(graph_changed=True)
-    def clear_tag(self, name: Name | Names, tag: str | Iterable[str]) -> None:
-        """Clear tag on a node or nodes. Ignored if tags are not set.
-
-        :param name: Node or nodes to clear tags for
-        :param tag: Tag to clear
-        """
-        apply_n(self._clear_tag_one, name, tag)
-
-    def _set_style_one(self, name: Name, style: str) -> None:
-        """Set style on a single node."""
-        node_key = to_nodekey(name)
-        self.dag.nodes[node_key][NodeAttributes.STYLE] = style
-
-    @_notifies_subscribers(graph_changed=True)
-    def set_style(self, name: Name | Names, style: str) -> None:
-        """Set styles on a node or nodes.
-
-        :param name: Node or nodes to set style for
-        :param style: Style to set
-        """
-        apply_n(self._set_style_one, name, style)
-
-    def _clear_style_one(self, name: Name) -> None:
-        """Clear style from a single node."""
-        node_key = to_nodekey(name)
-        self.dag.nodes[node_key][NodeAttributes.STYLE] = None
-
-    @_notifies_subscribers(graph_changed=True)
-    def clear_style(self, name: Name | Names) -> None:
-        """Clear style on a node or nodes.
-
-        :param name: Node or nodes to clear styles for
-        """
-        apply_n(self._clear_style_one, name)
-
-    def metadata(self, name: Name) -> dict[str, Any]:
-        """Get metadata for a node."""
-        node_key = to_nodekey(name)
-        if self.tree_has_path(name):
-            if node_key not in self._metadata:
-                self._metadata[node_key] = {}
-            result: dict[str, Any] = self._metadata[node_key]
-            return result
-        else:
-            msg = f"Node {node_key} does not exist."
-            raise NonExistentNodeException(msg)
-
-    @_notifies_subscribers(graph_changed=True)
-    def delete_node(self, name: Name) -> None:
-        """Delete a node from a computation.
-
-        When nodes are explicitly deleted with ``delete_node``, but are still depended on by other nodes, then they
-        will be set to PLACEHOLDER status. In this case, if the nodes that depend on a PLACEHOLDER node are deleted,
-        then the PLACEHOLDER node will also be deleted.
-
-        :param name: Name of the node to delete. If the node does not exist, a ``NonExistentNodeException`` will
-            be raised.
-        """
-        node_key = to_nodekey(name)
-        LOG.debug(f"Deleting node {node_key}")
-
-        if not self.dag.has_node(node_key):
-            msg = f"Node {node_key} does not exist"
-            raise NonExistentNodeException(msg)
-
-        if node_key in self._metadata:
-            del self._metadata[node_key]
-
-        if len(self.dag.succ[node_key]) == 0:
-            preds = self.dag.predecessors(node_key)
-            state = self.dag.nodes[node_key][NodeAttributes.STATE]
-            self.dag.remove_node(node_key)
-            self._state_map[state].remove(node_key)
-            self._mark_changed(node_key)
-            for n in preds:
-                if self.dag.nodes[n][NodeAttributes.STATE] == States.PLACEHOLDER:
-                    self.delete_node(n)
-        else:
-            self._set_state(node_key, States.PLACEHOLDER)
-
-    @_notifies_subscribers(graph_changed=True)
-    def rename_node(self, old_name: Name | Mapping[Name, Name], new_name: Name | None = None) -> None:
-        """Rename a node in a computation.
-
-        :param old_name: Node to rename, or a dictionary of nodes to rename, with existing names as keys, and
-            new names as values
-        :param new_name: New name for node.
-        """
-        name_mapping: dict[Name, Name]
-        if isinstance(old_name, Mapping) and not isinstance(old_name, str):
-            for k, v in old_name.items():
-                LOG.debug(f"Renaming node {k} to {v}")
-            if new_name is not None:
-                msg = "new_name must not be set if rename_node is passed a dictionary"
-                raise ValueError(msg)
-            else:
-                name_mapping = dict(old_name)  # type: ignore[arg-type]
-        else:
-            LOG.debug(f"Renaming node {old_name} to {new_name}")
-            old_node_key = to_nodekey(old_name)
-            if not self.dag.has_node(old_node_key):
-                msg = f"Node {old_name} does not exist"
-                raise NonExistentNodeException(msg)
-            assert new_name is not None  # noqa: S101
-            new_node_key = to_nodekey(new_name)
-            if self.dag.has_node(new_node_key):
-                msg = f"Node {new_name} already exists"
-                raise NodeAlreadyExistsException(msg)
-            name_mapping = {old_name: new_name}
-
-        node_key_mapping = {to_nodekey(on): to_nodekey(nn) for on, nn in name_mapping.items()}
-        nx.relabel_nodes(self.dag, node_key_mapping, copy=False)
-
-        for old_nk, new_nk in node_key_mapping.items():
-            if old_nk in self._metadata:
-                self._metadata[new_nk] = self._metadata[old_nk]
-                del self._metadata[old_nk]
-            else:
-                if new_nk in self._metadata:  # pragma: no cover
-                    del self._metadata[new_nk]
-
-        self._mark_changed(*node_key_mapping.keys(), *node_key_mapping.values())
-        self._refresh_maps()
-
-    @_notifies_subscribers(graph_changed=True)
-    def repoint(self, old_name: Name, new_name: Name) -> None:
-        """Changes all nodes that use old_name as an input to use new_name instead.
-
-        Note that if old_name is an input to new_name, then that will not be changed, to try to avoid introducing
-        circular dependencies, but other circular dependencies will not be checked.
-
-        If new_name does not exist, then it will be created as a PLACEHOLDER node.
-
-        :param old_name:
-        :param new_name:
-        :return:
-        """
-        old_node_key = to_nodekey(old_name)
-        new_node_key = to_nodekey(new_name)
-        if old_node_key == new_node_key:
-            return
-
-        changed_names = list(self.dag.successors(old_node_key))
-
-        if len(changed_names) > 0 and not self.dag.has_node(new_node_key):
-            self.dag.add_node(new_node_key, **{NodeAttributes.STATE: States.PLACEHOLDER})
-            self._state_map[States.PLACEHOLDER].add(new_node_key)
-
-        for name in changed_names:
-            if name == new_node_key:
-                continue
-            edge_data = self.dag.get_edge_data(old_node_key, name)
-            self.dag.add_edge(new_node_key, name, **edge_data)
-            self.dag.remove_edge(old_node_key, name)
-
-        for n in changed_names:
-            self.set_stale(n)
-
-    @_notifies_subscribers()
-    def insert(self, name: Name, value: Any, force: bool = False) -> None:
-        """Insert a value into a node of a computation.
-
-        Following insertation, the node will have state UPTODATE, and all its descendents will be COMPUTABLE or STALE.
-
-        If an attempt is made to insert a value into a node that does not exist, a ``NonExistentNodeException``
-        will be raised.
-
-        :param name: Name of the node to add.
-        :param value: The value to be inserted into the node.
-        :param force: Whether to force recalculation of descendents if node value and state would not be changed
-        """
-        node_key = to_nodekey(name)
-        LOG.debug(f"Inserting value into node {node_key}")
-
-        if not self.dag.has_node(node_key):
-            msg = f"Node {node_key} does not exist"
-            raise NonExistentNodeException(msg)
-
-        state = self._state_one(name)
-        if state == States.PLACEHOLDER:
-            msg = "Cannot insert into placeholder node. Use add_node to create the node first"
-            raise CannotInsertToPlaceholderNodeException(msg)
-
-        if not force and state == States.UPTODATE:
-            current_value = self._value_one(name)
-            if value_eq(value, current_value):
-                return
-
-        self._set_state_and_value(node_key, States.UPTODATE, value)
-        self._set_descendents(node_key, States.STALE)
-        for n in self.dag.successors(node_key):
-            self._try_set_computable(n)
-
-    @_notifies_subscribers()
-    def insert_many(self, name_value_pairs: Iterable[tuple[Name, object]]) -> None:
-        """Insert values into many nodes of a computation simultaneously.
-
-        Following insertation, the nodes will have state UPTODATE, and all their descendents will be COMPUTABLE
-        or STALE. In the case of inserting many nodes, some of which are descendents of others, this ensures that
-        the inserted nodes have correct status, rather than being set as STALE when their ancestors are inserted.
-
-        If an attempt is made to insert a value into a node that does not exist, a ``NonExistentNodeException`` will be
-        raised, and none of the nodes will be inserted.
-
-        :param name_value_pairs: Each tuple should be a pair (name, value), where name is the name of the node to
-            insert the value into.
-        :type name_value_pairs: List of tuples
-        """
-        node_key_value_pairs = [(to_nodekey(name), value) for name, value in name_value_pairs]
-        LOG.debug(f"Inserting value into nodes {', '.join(str(name) for name, value in node_key_value_pairs)}")
-
-        for name, _value in node_key_value_pairs:
-            if not self.dag.has_node(name):
-                msg = f"Node {name} does not exist"
-                raise NonExistentNodeException(msg)
-
-        stale = set()
-        computable = set()
-        for name, value in node_key_value_pairs:
-            self._set_state_and_value(name, States.UPTODATE, value)
-            stale.update(nx.dag.descendants(self.dag, name))
-            computable.update(self.dag.successors(name))
-        names = {name for name, value in node_key_value_pairs}
-        stale.difference_update(names)
-        computable.difference_update(names)
-        for name in stale:
-            self._set_state(name, States.STALE)
-        for name in computable:
-            self._try_set_computable(name)
-
-    @_notifies_subscribers()
-    def insert_from(self, other: "Computation", nodes: Iterable[Name] | None = None) -> None:
-        """Insert values into another Computation object into this Computation object.
-
-        :param other: The computation object to take values from
-        :type Computation:
-        :param nodes: Only populate the nodes with the names provided in this list. By default, all nodes from the
-            other Computation object that have corresponding nodes in this Computation object will be inserted
-        :type nodes: List, default None
-        """
-        if nodes is None:
-            nodes_set: set[Any] = set(self.dag.nodes)
-            nodes_set.intersection_update(other.dag.nodes())
-            nodes = nodes_set
-        name_value_pairs = [(name, other.value(name)) for name in nodes]
-        self.insert_many(name_value_pairs)
-
-    def _set_state(self, node_key: NodeKey, state: States) -> None:
-        """Set the state of a node without changing its value."""
-        node = self.dag.nodes[node_key]
-        old_state = node[NodeAttributes.STATE]
-        self._state_map[old_state].remove(node_key)
-        node[NodeAttributes.STATE] = state
-        self._state_map[state].add(node_key)
-        self._mark_changed(node_key)
-
-    def _set_state_and_value(
-        self, node_key: NodeKey, state: States, value: object, *, throw_conversion_exception: bool = True
-    ) -> None:
-        """Set state and value of a node, applying any converter."""
-        node = self.dag.nodes[node_key]
-        converter = node.get(NodeAttributes.CONVERTER)
-        if converter is None:
-            self._set_state_and_literal_value(node_key, state, value)
-        else:
-            try:
-                converted_value = converter(value)
-                self._set_state_and_literal_value(node_key, state, converted_value)
-            except Exception as e:
-                tb = traceback.format_exc()
-                self._set_error(node_key, e, tb)
-                if throw_conversion_exception:
-                    raise
-
-    def _set_state_and_literal_value(
-        self, node_key: NodeKey, state: States, value: object, require_old_state: bool = True
-    ) -> None:
-        """Set state and literal value of a node without conversion."""
-        node = self.dag.nodes[node_key]
-        try:
-            old_state = node[NodeAttributes.STATE]
-            self._state_map[old_state].remove(node_key)
-        except KeyError:
-            if require_old_state:
-                raise  # pragma: no cover
-        node[NodeAttributes.STATE] = state
-        node[NodeAttributes.VALUE] = value
-        self._state_map[state].add(node_key)
-        self._mark_changed(node_key)
-
-    def _set_states(self, node_keys: Iterable[NodeKey], state: States) -> None:
-        """Set the state of multiple nodes at once.
-
-        Materialising the keys is only needed when they have to be read a
-        second time, which is only when a subscriber will be told about them.
-        Doing it unconditionally turned the set that callers pass into a tuple,
-        and ``set.update(tuple)`` rehashes every element where
-        ``set.update(set)`` reuses the hashes the source set already holds ---
-        399 extra hashes per insert on a 400-node chain, and the whole of the
-        subscription API's measured cost to callers who never subscribe.
-        """
-        watched = bool(self._subscriptions)
-        if watched:
-            node_keys = tuple(node_keys)
-        for name in node_keys:
-            node = self.dag.nodes[name]
-            old_state = node[NodeAttributes.STATE]
-            self._state_map[old_state].remove(name)
-            node[NodeAttributes.STATE] = state
-        self._state_map[state].update(node_keys)
-        if watched:
-            # Not via _mark_changed: its star-args would repack the keys again.
-            self._pending_changed_nodes.update(node_keys)
-
-    @_notifies_subscribers()
-    def set_stale(self, name: Name) -> None:
-        """Set the state of a node and all its dependencies to STALE.
-
-        :param name: Name of the node to set as STALE.
-        """
-        node_key = to_nodekey(name)
-        node_keys: list[NodeKey] = [node_key]
-        node_keys.extend(nx.dag.descendants(self.dag, node_key))
-        self._set_states(node_keys, States.STALE)
-        self._try_set_computable(node_key)
-
-    @_notifies_subscribers()
-    def pin(self, name: Name, value: Any = None) -> None:
-        """Set the state of a node to PINNED.
-
-        :param name: Name of the node to set as PINNED.
-        :param value: Value to pin to the node, if provided.
-        :type value: default None
-        """
-        node_key = to_nodekey(name)
-        if value is not None:
-            self.insert(node_key, value)
-        self._set_states([node_key], States.PINNED)
-
-    @_notifies_subscribers()
-    def unpin(self, name: Name) -> None:
-        """Unpin a node (state of node and all descendents will be set to STALE).
-
-        :param name: Name of the node to set as PINNED.
-        """
-        node_key = to_nodekey(name)
-        self.set_stale(node_key)
-
-    def _get_descendents(self, node_key: NodeKey, stop_states: set[States] | None = None) -> set[NodeKey]:
-        """Get all descendant nodes, optionally stopping at certain states."""
-        if stop_states is None:
-            stop_states = set()
-        if self.dag.nodes[node_key][NodeAttributes.STATE] in stop_states:
-            return set()
-        visited = set()
-        to_visit = {node_key}
-        while to_visit:
-            n = to_visit.pop()
-            visited.add(n)
-            for n1 in self.dag.successors(n):
-                if n1 in visited:
-                    continue
-                if self.dag.nodes[n1][NodeAttributes.STATE] in stop_states:
-                    continue
-                to_visit.add(n1)
-        visited.remove(node_key)
-        return visited
-
-    def _set_descendents(self, node_key: NodeKey, state: States) -> None:
-        """Set the state of all descendant nodes."""
-        descendents = self._get_descendents(node_key, {States.PINNED})
-        self._set_states(descendents, state)
-
-    def _set_uninitialized(self, node_key: NodeKey) -> None:
-        """Set a node to uninitialized state and clear its value."""
-        self._set_states([node_key], States.UNINITIALIZED)
-        self.dag.nodes[node_key].pop(NodeAttributes.VALUE, None)
-
-    def _set_uptodate(self, node_key: NodeKey, value: object) -> None:
-        """Set a node to up-to-date state with a value."""
-        self._set_state_and_value(node_key, States.UPTODATE, value)
-        self._set_descendents(node_key, States.STALE)
-        for n in self.dag.successors(node_key):
-            self._try_set_computable(n)
-
-    def _set_error(self, node_key: NodeKey, exc: Exception, tb: str) -> None:
-        """Set a node to error state with exception information."""
-        self._set_state_and_literal_value(node_key, States.ERROR, Error(exc, tb))
-        self._set_descendents(node_key, States.STALE)
-
-    def _try_set_computable(self, node_key: NodeKey) -> None:
-        """Set node to computable if all predecessors are up-to-date."""
-        if self.dag.nodes[node_key][NodeAttributes.STATE] == States.PINNED:
-            return
-        if self.dag.nodes[node_key].get(NodeAttributes.FUNC) is not None:
-            for n in self.dag.predecessors(node_key):
-                if not self.dag.has_node(n):
-                    return  # pragma: no cover
-                if self.dag.nodes[n][NodeAttributes.STATE] not in (States.UPTODATE, States.PINNED):
-                    return
-            self._set_state(node_key, States.COMPUTABLE)
-
     def validate(self) -> ValidationReport:
         """Inspect the entire graph for structural and readiness problems.
 
@@ -1282,993 +390,6 @@ class Computation:
                     msg = f"Node {node_key} does not exist"
                     raise NonExistentNodeException(msg)
         return create_execution_plan(self.dag, self.executor_map, target_node_keys)
-
-    def _get_parameter_data(self, node_key: NodeKey) -> Iterable[_ParameterItem]:
-        """Get all parameter data for a node's function call."""
-        for arg, value in self.dag.nodes[node_key][NodeAttributes.ARGS].items():
-            yield _ParameterItem(_ParameterType.ARG, arg, value)
-        for param_name, value in self.dag.nodes[node_key][NodeAttributes.KWDS].items():
-            yield _ParameterItem(_ParameterType.KWD, param_name, value)
-        for in_node_name in self.dag.predecessors(node_key):
-            param_value = self.dag.nodes[in_node_name][NodeAttributes.VALUE]
-            edge = self.dag[in_node_name][node_key]
-            param_type, param_name = edge[EdgeAttributes.PARAM]
-            yield _ParameterItem(param_type, param_name, param_value)
-
-    def _get_func_args_kwds(
-        self, node_key: NodeKey
-    ) -> tuple[Callable[..., Any], str | None, list[Any], dict[str, Any]]:
-        """Get function, executor name, args and kwargs for a node."""
-        node0 = self.dag.nodes[node_key]
-        f = node0[NodeAttributes.FUNC]
-        executor_name = node0.get(NodeAttributes.EXECUTOR)
-        args: list[Any] = []
-        kwds: dict[str, Any] = {}
-        for param in self._get_parameter_data(node_key):
-            if param.type == _ParameterType.ARG:
-                idx = param.name
-                assert isinstance(idx, int)  # noqa: S101
-                while len(args) <= idx:
-                    args.append(None)
-                args[idx] = param.value
-            elif param.type == _ParameterType.KWD:
-                assert isinstance(param.name, str)  # noqa: S101
-                kwds[param.name] = param.value
-            else:  # pragma: no cover
-                msg = f"Unexpected param type: {param.type}"
-                raise ValidationError(msg)
-        return f, executor_name, args, kwds
-
-    def get_definition_args_kwds(self, name: Name) -> tuple[list[Any], dict[str, Any]]:
-        """Get the arguments and keyword arguments for a node's function definition."""
-        res_args: list[Any] = []
-        res_kwds: dict[str, Any] = {}
-        node_key = to_nodekey(name)
-        node_data = self.dag.nodes[node_key]
-        if NodeAttributes.ARGS in node_data:
-            for idx, value in node_data[NodeAttributes.ARGS].items():
-                while len(res_args) <= idx:
-                    res_args.append(None)
-                res_args[idx] = C(value)
-        if NodeAttributes.KWDS in node_data:
-            for param_name, value in node_data[NodeAttributes.KWDS].items():
-                res_kwds[param_name] = C(value)
-        for in_node_key in self.dag.predecessors(node_key):
-            edge = self.dag[in_node_key][node_key]
-            if EdgeAttributes.PARAM in edge:
-                param_type, param_name = edge[EdgeAttributes.PARAM]
-                if param_type == _ParameterType.ARG:
-                    idx = param_name
-                    assert isinstance(idx, int)  # noqa: S101
-                    while len(res_args) <= idx:
-                        res_args.append(None)
-                    res_args[idx] = in_node_key.name
-                elif param_type == _ParameterType.KWD:
-                    res_kwds[param_name] = in_node_key.name
-                else:  # pragma: no cover
-                    msg = f"Unexpected param type: {param_type}"
-                    raise ValidationError(msg)
-        return res_args, res_kwds
-
-    def _compute_nodes(self, node_keys: Iterable[NodeKey], raise_exceptions: bool = False) -> None:
-        """Compute multiple nodes, handling dependencies and parallel execution."""
-        LOG.debug(f"Computing nodes {node_keys}")
-
-        futs: dict[Any, NodeKey] = {}
-        node_keys_set = set(node_keys)
-
-        def run(name: NodeKey) -> None:
-            """Submit a node computation to an executor."""
-            f, executor_name, args, kwds = self._get_func_args_kwds(name)
-            executor = self.default_executor if executor_name is None else self.executor_map[executor_name]
-            fut = executor.submit(_eval_node, name, f, args, kwds, raise_exceptions)
-            futs[fut] = name
-
-        computed: set[NodeKey] = set()
-
-        for node_key in node_keys_set:
-            node0 = self.dag.nodes[node_key]
-            state = node0[NodeAttributes.STATE]
-            if state == States.COMPUTABLE:
-                run(node_key)
-
-        while len(futs) > 0:
-            done, _not_done = wait(futs.keys(), return_when=FIRST_COMPLETED)
-            for fut in done:
-                node_key = futs.pop(fut)
-                node0 = self.dag.nodes[node_key]
-                try:
-                    value, exc, tb, start_dt, end_dt = fut.result()
-                except Exception as e:
-                    exc = e
-                    tb = traceback.format_exc()
-                    self._set_error(node_key, exc, tb)
-                    raise
-                delta = (end_dt - start_dt).total_seconds()
-                if exc is None:
-                    self._set_state_and_value(node_key, States.UPTODATE, value, throw_conversion_exception=False)
-                    node0[NodeAttributes.TIMING] = TimingData(start_dt, end_dt, delta)
-                    self._set_descendents(node_key, States.STALE)
-                    for n in self.dag.successors(node_key):
-                        logging.debug(str(node_key) + " " + str(n) + " " + str(computed))
-                        if n in computed:
-                            msg = f"Calculating {node_key} for the second time"
-                            raise LoopDetectedException(msg)
-                        self._try_set_computable(n)
-                        node0 = self.dag.nodes[n]
-                        state = node0[NodeAttributes.STATE]
-                        if state == States.COMPUTABLE and n in node_keys_set:
-                            run(n)
-                else:
-                    assert tb is not None  # noqa: S101
-                    self._set_error(node_key, exc, tb)
-                computed.add(node_key)
-
-    def _get_calc_node_keys(self, node_key: NodeKey) -> list[NodeKey]:
-        """Get node keys that need to be computed for a target node."""
-        g = nx.DiGraph()
-        g.add_nodes_from(self.dag.nodes)
-        g.add_edges_from(self.dag.edges)
-        for n in nx.ancestors(g, node_key):
-            node = self.dag.nodes[n]
-            state = node[NodeAttributes.STATE]
-            if state == States.UPTODATE or state == States.PINNED:
-                g.remove_node(n)
-
-        ancestors = nx.ancestors(g, node_key)
-        for n in ancestors:
-            node = self.dag.nodes[n]
-            state = node[NodeAttributes.STATE]
-
-            if state == States.UNINITIALIZED and len(self.dag.pred[n]) == 0:
-                msg = f"Cannot compute {node_key} because {n} uninitialized"
-                raise ValidationError(msg)
-            if state == States.PLACEHOLDER:
-                msg = f"Cannot compute {node_key} because {n} is placeholder"
-                raise ValidationError(msg)
-
-        ancestors.add(node_key)
-        g = g.subgraph(ancestors).copy()
-        nodes_sorted = topological_sort(g)
-        return list(nodes_sorted)
-
-    def _get_calc_node_names(self, name: Name) -> Names:
-        """Get node names that need to be computed for a target node."""
-        node_key = to_nodekey(name)
-        return node_keys_to_names(self._get_calc_node_keys(node_key))
-
-    @_notifies_subscribers()
-    def compute(self, name: Name | Iterable[Name], raise_exceptions: bool = False) -> None:
-        """Compute a node or block and all necessary predecessors.
-
-        Following the computation, if successful, the target node, and all necessary ancestors that were not already
-        UPTODATE will have been calculated and set to UPTODATE. Any node that did not need to be calculated will not
-        have been recalculated.
-
-        If any nodes raises an exception, then the state of that node will be set to ERROR, and its value set to an
-        object containing the exception object, as well as a traceback. This will not halt the computation, which
-        will proceed as far as it can, until no more nodes that would be required to calculate the target are
-        COMPUTABLE.
-
-        A block name computes every node below that path. Multiple node and block
-        names may be supplied in a list or generator.
-
-        :param name: Name of the node or block to compute
-        :param raise_exceptions: Whether to pass exceptions raised by node computations back to the caller
-        :type raise_exceptions: Boolean, default False
-        """
-        calc_nodes: set[NodeKey] = set()
-        names = name if isinstance(name, (types.GeneratorType, list)) else [name]
-        for name0 in names:
-            node_key = to_nodekey(name0)
-            targets = (
-                [node_key]
-                if self.has_node(node_key)
-                else names_to_node_keys(self.get_tree_descendents(node_key, graph_nodes_only=True))
-            )
-            if not targets:
-                targets = [node_key]
-            for target in targets:
-                calc_nodes.update(self._get_calc_node_keys(target))
-        self._compute_nodes(calc_nodes, raise_exceptions=raise_exceptions)
-
-    @_notifies_subscribers()
-    def compute_all(self, raise_exceptions: bool = False) -> None:
-        """Compute all nodes of a computation that can be computed.
-
-        Nodes that are already UPTODATE will not be recalculated. Following the computation, if successful, all
-        nodes will have state UPTODATE, except UNINITIALIZED input nodes and PLACEHOLDER nodes.
-
-        If any nodes raises an exception, then the state of that node will be set to ERROR, and its value set to an
-        object containing the exception object, as well as a traceback. This will not halt the computation, which
-        will proceed as far as it can, until no more nodes are COMPUTABLE.
-
-        :param raise_exceptions: Whether to pass exceptions raised by node computations back to the caller
-        :type raise_exceptions: Boolean, default False
-        """
-        self._compute_nodes(self._node_keys(), raise_exceptions=raise_exceptions)
-
-    def _node_keys(self) -> list[NodeKey]:
-        """Get a list of nodes in this computation.
-
-        :return: List of nodes.
-        """
-        return list(self.dag.nodes)
-
-    def nodes(self) -> list[Name]:
-        """Get a list of nodes in this computation.
-
-        :return: List of nodes.
-        """
-        return [n.name for n in self.dag.nodes]
-
-    def get_tree_list_children(self, name: Name) -> set[Name]:
-        """Get a list of nodes in this computation.
-
-        :return: List of nodes.
-        """
-        node_key = to_nodekey(name)
-        idx = len(node_key.parts)
-        result = set()
-        for n in self.dag.nodes:
-            if n.is_descendent_of(node_key):
-                result.add(n.parts[idx])
-        return result
-
-    def has_node(self, name: Name) -> bool:
-        """Check if a node with the given name exists in the computation."""
-        node_key = to_nodekey(name)
-        return node_key in self.dag.nodes
-
-    def tree_has_path(self, name: Name) -> bool:
-        """Check if a hierarchical path exists in the computation tree."""
-        node_key = to_nodekey(name)
-        if node_key.is_root:
-            return True
-        if self.has_node(node_key):
-            return True
-        return any(n.is_descendent_of(node_key) for n in self.dag.nodes)
-
-    def get_tree_descendents(
-        self, name: Name | None = None, *, include_stem: bool = True, graph_nodes_only: bool = False
-    ) -> set[Name]:
-        """Get a list of descendent blocks and nodes.
-
-        Returns blocks and nodes that are descendents of the input node,
-        e.g. for node 'foo', might return ['foo/bar', 'foo/baz'].
-
-        :param name: Name of node to get descendents for
-        :return: List of descendent node names
-        """
-        node_key = NodeKey.root() if name is None else to_nodekey(name)
-        stemsize = len(node_key.parts)
-        result = set()
-        for n in self.dag.nodes:
-            if n.is_descendent_of(node_key):
-                nodes = [n] if graph_nodes_only else n.ancestors()
-                for n2 in nodes:
-                    if n2.is_descendent_of(node_key):
-                        nm = n2.name if include_stem else NodeKey(tuple(n2.parts[stemsize:])).name
-                        result.add(nm)
-        return result
-
-    def _state_one(self, name: Name) -> States:
-        """Get the state of a single node."""
-        node_key = to_nodekey(name)
-        state: States = self.dag.nodes[node_key][NodeAttributes.STATE]
-        return state
-
-    @overload
-    def state(self, name: Name) -> States: ...
-
-    @overload
-    def state(self, name: Names) -> list[States]: ...
-
-    def state(self, name: Name | Names) -> States | list[States]:
-        """Get the state of a node.
-
-        This can also be accessed using the attribute-style accessor ``s`` if ``name`` is a valid Python
-        attribute name::
-
-            >>> comp = Computation()
-            >>> comp.add_node('foo', value=1)
-            >>> comp.state('foo')
-            <States.UPTODATE: 4>
-            >>> comp.s.foo
-            <States.UPTODATE: 4>
-
-        :param name: Name or names of the node to get state for
-        :type name: Name or Names
-        """
-        return apply1(self._state_one, name)
-
-    def _value_one(self, name: Name) -> Any:
-        """Get the value of a single node."""
-        node_key = to_nodekey(name)
-        return self.dag.nodes[node_key][NodeAttributes.VALUE]
-
-    @overload
-    def value(self, name: Name) -> Any: ...
-
-    @overload
-    def value(self, name: Names) -> list[Any]: ...
-
-    def value(self, name: Name | Names) -> Any | list[Any]:
-        """Get the current value of a node.
-
-        This can also be accessed using the attribute-style accessor ``v`` if ``name`` is a valid Python
-        attribute name::
-
-            >>> comp = Computation()
-            >>> comp.add_node('foo', value=1)
-            >>> comp.value('foo')
-            1
-            >>> comp.v.foo
-            1
-
-        :param name: Name or names of the node to get the value of
-        :type name: Name or Names
-        """
-        return apply1(self._value_one, name)
-
-    def compute_and_get_value(self, name: Name) -> Any:
-        """Get the current value of a node.
-
-        This can also be accessed using the attribute-style accessor ``v`` if ``name`` is a valid Python
-        attribute name::
-
-            >>> comp = Computation()
-            >>> comp.add_node('foo', value=1)
-            >>> comp.add_node('bar', lambda foo: foo + 1)
-            >>> comp.compute_and_get_value('bar')
-            2
-            >>> comp.x.bar
-            2
-
-        :param name: Name or names of the node to get the value of
-        :type name: Name
-        """
-        nk = to_nodekey(name)
-        if self.state(nk) == States.UPTODATE:
-            return self.value(nk)
-        self.compute(nk, raise_exceptions=True)
-        if self.state(nk) == States.UPTODATE:
-            return self.value(nk)
-        msg = f"Unable to compute node {nk}"
-        raise ComputationError(msg)
-
-    def _tag_one(self, name: Name) -> set[str]:
-        """Get the tags of a single node."""
-        node_key = to_nodekey(name)
-        node = self.dag.nodes[node_key]
-        tags: set[str] = node[NodeAttributes.TAG]
-        return tags
-
-    @overload
-    def tags(self, name: Name) -> set[str]: ...
-
-    @overload
-    def tags(self, name: Names) -> list[set[str]]: ...
-
-    def tags(self, name: Name | Names) -> set[str] | list[set[str]]:
-        """Get the tags associated with a node.
-
-            >>> comp = Computation()
-            >>> comp.add_node('a', tags=['foo', 'bar'])
-            >>> sorted(comp.t.a)
-            ['__serialize__', 'bar', 'foo']
-
-        :param name: Name or names of the node to get the tags of
-        :return:
-        """
-        return apply1(self._tag_one, name)
-
-    def nodes_by_tag(self, tag: str | Iterable[str]) -> set[Name]:
-        """Get the names of nodes with a particular tag or tags.
-
-        :param tag: Tag or tags for which to retrieve nodes
-        :return: Names of the nodes with those tags
-        """
-        nodes: set[NodeKey] = set()
-        tags_to_check: Iterable[str] = [tag] if isinstance(tag, str) else tag
-        for tag1 in tags_to_check:
-            nodes1 = self._tag_map.get(tag1)
-            if nodes1 is not None:
-                nodes.update(nodes1)
-        return {n.name for n in nodes}
-
-    def _style_one(self, name: Name) -> str | None:
-        """Get the style of a single node."""
-        node_key = to_nodekey(name)
-        node = self.dag.nodes[node_key]
-        style: str | None = node.get(NodeAttributes.STYLE)
-        return style
-
-    @overload
-    def styles(self, name: Name) -> str | None: ...
-
-    @overload
-    def styles(self, name: Names) -> list[str | None]: ...
-
-    def styles(self, name: Name | Names) -> str | None | list[str | None]:
-        """Get the tags associated with a node.
-
-            >>> comp = Computation()
-            >>> comp.add_node('a', style='dot')
-            >>> comp.style.a
-            'dot'
-
-        :param name: Name or names of the node to get the tags of
-        :return:
-        """
-        return apply1(self._style_one, name)
-
-    def _get_item_one(self, name: Name) -> NodeData:
-        """Get state and value data for a single node."""
-        node_key = to_nodekey(name)
-        node = self.dag.nodes[node_key]
-        return NodeData(node[NodeAttributes.STATE], node[NodeAttributes.VALUE])
-
-    @overload
-    def __getitem__(self, name: Name) -> NodeData: ...
-
-    @overload
-    def __getitem__(self, name: Names) -> list[NodeData]: ...
-
-    def __getitem__(self, name: Name | Names) -> NodeData | list[NodeData]:
-        """Get the state and current value of a node.
-
-        :param name: Name of the node to get the state and value of
-        """
-        return apply1(self._get_item_one, name)
-
-    def _get_timing_one(self, name: Name) -> TimingData | None:
-        """Get timing data for a single node."""
-        node_key = to_nodekey(name)
-        node = self.dag.nodes[node_key]
-        timing: TimingData | None = node.get(NodeAttributes.TIMING, None)
-        return timing
-
-    @overload
-    def get_timing(self, name: Name) -> TimingData | None: ...
-
-    @overload
-    def get_timing(self, name: Names) -> list[TimingData | None]: ...
-
-    def get_timing(self, name: Name | Names) -> TimingData | None | list[TimingData | None]:
-        """Get the timing information for a node.
-
-        :param name: Name or names of the node to get the timing information of
-        :return:
-        """
-        return apply1(self._get_timing_one, name)
-
-    def to_df(self) -> pd.DataFrame:
-        """Get a dataframe containing the states and value of all nodes of computation.
-
-        ::
-
-            >>> import loman
-            >>> comp = loman.Computation()
-            >>> comp.add_node('foo', value=1)
-            >>> comp.add_node('bar', value=2)
-            >>> comp.to_df()  # doctest: +NORMALIZE_WHITESPACE
-                           state  value
-            foo  States.UPTODATE      1
-            bar  States.UPTODATE      2
-        """
-        df = pd.DataFrame(index=topological_sort(self.dag))
-        df[NodeAttributes.STATE] = pd.Series(nx.get_node_attributes(self.dag, NodeAttributes.STATE))
-        df[NodeAttributes.VALUE] = pd.Series(nx.get_node_attributes(self.dag, NodeAttributes.VALUE))
-        df_timing = pd.DataFrame.from_dict(nx.get_node_attributes(self.dag, "timing"), orient="index")
-        df = pd.merge(df, df_timing, left_index=True, right_index=True, how="left")
-        df.index = pd.Index([nk.name for nk in df.index])
-        return df
-
-    def to_dict(self) -> dict[NodeKey, Any]:
-        """Get a dictionary containing the values of all nodes of a computation.
-
-        ::
-
-            >>> import loman
-            >>> comp = loman.Computation()
-            >>> comp.add_node('foo', value=1)
-            >>> comp.add_node('bar', value=2)
-            >>> comp.to_dict()  # doctest: +ELLIPSIS
-            {NodeKey('foo'): 1, NodeKey('bar'): 2}
-        """
-        result: dict[NodeKey, Any] = nx.get_node_attributes(self.dag, NodeAttributes.VALUE)
-        return result
-
-    def _get_inputs_one_node_keys(self, node_key: NodeKey) -> list[NodeKey | None]:
-        """Get input node keys for a single node."""
-        args_dict: dict[int, NodeKey] = {}
-        kwds: list[NodeKey | None] = []
-        max_arg_index = -1
-        for input_node in self.dag.predecessors(node_key):
-            input_edge = self.dag[input_node][node_key]
-            input_type, input_param = input_edge[EdgeAttributes.PARAM]
-            if input_type == _ParameterType.ARG:
-                idx = input_param
-                max_arg_index = max(max_arg_index, idx)
-                args_dict[idx] = input_node
-            elif input_type == _ParameterType.KWD:
-                kwds.append(input_node)
-        if max_arg_index >= 0:
-            args: list[NodeKey | None] = [None] * (max_arg_index + 1)
-            for idx, input_node in args_dict.items():
-                args[idx] = input_node
-            result: list[NodeKey | None] = args + kwds
-            return result
-        else:
-            return kwds
-
-    def _get_inputs_one_names(self, name: Name) -> Names:
-        """Get input node names for a single node."""
-        node_key = to_nodekey(name)
-        return node_keys_to_names([nk for nk in self._get_inputs_one_node_keys(node_key) if nk is not None])
-
-    @overload
-    def get_inputs(self, name: Name) -> Names: ...
-
-    @overload
-    def get_inputs(self, name: Names) -> list[Names]: ...
-
-    def get_inputs(self, name: Name | Names) -> Names | list[Names]:
-        """Get a list of the inputs for a node or set of nodes.
-
-        :param name: Name or names of nodes to get inputs for
-        :return: If name is scalar, return a list of upstream nodes used as input. If name is a list, return a
-            list of list of inputs.
-        """
-        return apply1(self._get_inputs_one_names, name)
-
-    def _get_ancestors_node_keys(self, node_keys: Iterable[NodeKey], include_self: bool = True) -> set[NodeKey]:
-        """Get all ancestor node keys for a set of nodes."""
-        ancestors: set[NodeKey] = set()
-        for n in node_keys:
-            if include_self:
-                ancestors.add(n)
-            for ancestor in nx.ancestors(self.dag, n):
-                ancestors.add(ancestor)
-        return ancestors
-
-    def get_ancestors(self, names: Name | Names, include_self: bool = True) -> Names:
-        """Get all ancestor nodes of the specified nodes."""
-        node_keys = names_to_node_keys(names)
-        ancestor_node_keys = self._get_ancestors_node_keys(node_keys, include_self)
-        return node_keys_to_names(ancestor_node_keys)
-
-    def _get_original_inputs_node_keys(self, node_keys: list[NodeKey] | None) -> list[NodeKey]:
-        """Get original input node keys that have no computation function."""
-        resolved_node_keys: Iterable[NodeKey]
-        resolved_node_keys = self._node_keys() if node_keys is None else self._get_ancestors_node_keys(node_keys)
-        return [n for n in resolved_node_keys if self.dag.nodes[n].get(NodeAttributes.FUNC) is None]
-
-    def get_original_inputs(self, names: Name | Names | None = None) -> Names:
-        """Get a list of the original non-computed inputs for a node or set of nodes.
-
-        :param names: Name or names of nodes to get inputs for
-        :return: Return a list of original non-computed inputs that are ancestors of the input nodes
-        """
-        nks = None if names is None else names_to_node_keys(names)
-
-        result_nks = self._get_original_inputs_node_keys(nks)
-
-        return node_keys_to_names(result_nks)
-
-    def _get_outputs_one(self, name: Name) -> Names:
-        """Get output node names for a single node."""
-        node_key = to_nodekey(name)
-        output_node_keys = list(self.dag.successors(node_key))
-        return node_keys_to_names(output_node_keys)
-
-    @overload
-    def get_outputs(self, name: Name) -> Names: ...
-
-    @overload
-    def get_outputs(self, name: Names) -> list[Names]: ...
-
-    def get_outputs(self, name: Name | Names) -> Names | list[Names]:
-        """Get a list of the outputs for a node or set of nodes.
-
-        :param name: Name or names of nodes to get outputs for
-        :return: If name is scalar, return a list of downstream nodes used as output. If name is a list, return a
-            list of list of outputs.
-
-        """
-        return apply1(self._get_outputs_one, name)
-
-    def _get_descendents_node_keys(self, node_keys: Iterable[NodeKey], include_self: bool = True) -> set[NodeKey]:
-        """Get all descendant node keys for a set of nodes."""
-        descendent_node_keys: set[NodeKey] = set()
-        for node_key in node_keys:
-            if include_self:
-                descendent_node_keys.add(node_key)
-            for descendent in nx.descendants(self.dag, node_key):
-                descendent_node_keys.add(descendent)
-        return descendent_node_keys
-
-    def get_descendents(self, names: Name | Names, include_self: bool = True) -> Names:
-        """Get all descendent nodes of the specified nodes."""
-        node_keys = names_to_node_keys(names)
-        descendent_node_keys = self._get_descendents_node_keys(node_keys, include_self)
-        return node_keys_to_names(descendent_node_keys)
-
-    def get_final_outputs(self, names: Name | Names | None = None) -> Names:
-        """Get final output nodes (nodes with no descendants) from the specified nodes."""
-        final_node_keys: Iterable[NodeKey]
-        if names is None:
-            final_node_keys = self._node_keys()
-        else:
-            nks = names_to_node_keys(names)
-            final_node_keys = self._get_descendents_node_keys(nks)
-        output_node_keys = [n for n in final_node_keys if len(nx.descendants(self.dag, n)) == 0]
-        return node_keys_to_names(output_node_keys)
-
-    def get_source(self, name: Name) -> str:
-        """Get the source code for a node."""
-        node_key = to_nodekey(name)
-        func = self.dag.nodes[node_key].get(NodeAttributes.FUNC, None)
-        if func is not None:
-            file = inspect.getsourcefile(func)
-            _, lineno = inspect.getsourcelines(func)
-            source = inspect.getsource(func)
-            return f"{file}:{lineno}\n\n{source}"
-        else:
-            return "NOT A CALCULATED NODE"
-
-    def print_source(self, name: Name) -> None:
-        """Print the source code for a computation node."""
-        print(self.get_source(name))
-
-    @_notifies_subscribers(graph_changed=True)
-    def restrict(self, output_names: Name | Names, input_names: Name | Names | None = None) -> None:
-        """Restrict a computation to the ancestors of a set of output nodes.
-
-        Excludes ancestors of a set of input nodes.
-
-        If the set of input_nodes that is specified is not sufficient for the set of output_nodes then additional
-        nodes that are ancestors of the output_nodes will be included, but the input nodes specified will be input
-        nodes of the modified Computation.
-
-        :param output_nodes:
-        :param input_nodes:
-        :return: None - modifies existing computation in place
-        """
-        if input_names is not None:
-            for n in as_iterable(input_names):
-                nodedata = self._get_item_one(n)
-                self.add_node(n)
-                self._set_state_and_literal_value(to_nodekey(n), nodedata.state, nodedata.value)
-        output_node_keys = names_to_node_keys(output_names)
-        ancestor_node_keys = self._get_ancestors_node_keys(output_node_keys)
-        removed = [n for n in self.dag if n not in ancestor_node_keys]
-        self.dag.remove_nodes_from(removed)
-        self._mark_changed(*removed)
-
-    def __getstate__(self) -> dict[str, Any]:
-        """Prepare computation for serialization by removing non-serializable nodes."""
-        node_serialize = nx.get_node_attributes(self.dag, NodeAttributes.TAG)
-        obj = self.copy()
-        for name, tags in node_serialize.items():
-            if SystemTags.SERIALIZE not in tags:
-                obj._set_uninitialized(name)
-        return {"dag": obj.dag}
-
-    def __setstate__(self, state: dict[str, Any]) -> None:
-        """Restore computation from serialized state."""
-        self.__init__()
-        self.dag = state["dag"]
-        self._refresh_maps()
-
-    def write_dill_old(self, file_: str | BinaryIO) -> None:
-        """Serialize a computation to a file or file-like object.
-
-        :param file_: If string, writes to a file
-        :type file_: File-like object, or string
-        """
-        warnings.warn("write_dill_old is deprecated, use write_dill instead", DeprecationWarning, stacklevel=2)
-        original_getstate = self.__class__.__getstate__
-        original_setstate = self.__class__.__setstate__
-
-        try:
-            del self.__class__.__getstate__
-            del self.__class__.__setstate__
-
-            node_serialize = nx.get_node_attributes(self.dag, NodeAttributes.TAG)
-            obj = self.copy()
-            obj.executor_map = None  # type: ignore[assignment]
-            obj.default_executor = None  # type: ignore[assignment]
-            for name, tags in node_serialize.items():
-                if SystemTags.SERIALIZE not in tags:
-                    obj._set_uninitialized(name)
-
-            if isinstance(file_, str):
-                with open(file_, "wb") as f:
-                    dill.dump(obj, f)
-            else:
-                dill.dump(obj, file_)
-        finally:
-            self.__class__.__getstate__ = original_getstate  # type: ignore[method-assign]
-            self.__class__.__setstate__ = original_setstate
-
-    def write_dill(self, file_: str | BinaryIO) -> None:
-        """Serialize a computation to a file or file-like object.
-
-        .. deprecated::
-            Use :meth:`write_json` instead.  dill-based serialization will be
-            removed in a future release.
-
-        :param file_: If string, writes to a file
-        :type file_: File-like object, or string
-        """
-        warnings.warn(
-            "write_dill is deprecated and will be removed in a future release. Use write_json instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        if isinstance(file_, str):
-            with open(file_, "wb") as f:
-                dill.dump(self, f)
-        else:
-            dill.dump(self, file_)
-
-    @staticmethod
-    def read_dill(file_: str | BinaryIO) -> "Computation":
-        """Deserialize a computation from a file or file-like object.
-
-        .. deprecated::
-            Use :meth:`read_json` instead.  dill-based serialization will be
-            removed in a future release.
-
-        .. warning::
-            This method uses dill.load() which can execute arbitrary code.
-            Only load files from trusted sources. Never load data from
-            untrusted or unauthenticated sources as it may lead to arbitrary
-            code execution.
-
-        :param file_: If string, writes to a file
-        :type file_: File-like object, or string
-        """
-        warnings.warn(
-            "read_dill is deprecated and will be removed in a future release. Use read_json instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        if isinstance(file_, str):
-            with open(file_, "rb") as f:
-                obj = dill.load(f)  # noqa: S301  # nosec B301
-        else:
-            obj = dill.load(file_)  # noqa: S301  # nosec B301
-        if isinstance(obj, Computation):
-            return obj
-        else:
-            msg = "Loaded object is not a Computation"
-            raise ValidationError(msg)
-
-    def write_json(self, file_: str | TextIO, *, serializer: "ComputationSerializer | None" = None) -> None:
-        """Serialize a computation to a JSON file or file-like object.
-
-        Custom types can be supported by passing a custom *serializer* —
-        either a :class:`~loman.serialization.computation.ComputationSerializer`
-        instance with extra transformers registered, or a subclass that
-        overrides the transformer factory.
-
-        :param file_: Destination file path (str) or text-mode file-like object.
-        :param serializer: Optional custom serializer.  If ``None`` the default
-            :class:`~loman.serialization.computation.ComputationSerializer` is used.
-        """
-        from .serialization.computation import ComputationSerializer
-
-        s = serializer if serializer is not None else ComputationSerializer()
-        if isinstance(file_, str):
-            with open(file_, "w", encoding="utf-8") as f:
-                s.dump(self, f)
-        else:
-            s.dump(self, file_)
-
-    @staticmethod
-    def read_json(file_: str | TextIO, *, serializer: "ComputationSerializer | None" = None) -> "Computation":
-        """Deserialize a computation from a JSON file or file-like object.
-
-        :param file_: Source file path (str) or text-mode file-like object.
-        :param serializer: Optional custom serializer.  If ``None`` the default
-            :class:`~loman.serialization.computation.ComputationSerializer` is used.
-        :rtype: Computation
-        """
-        from .serialization.computation import ComputationSerializer
-
-        s = serializer if serializer is not None else ComputationSerializer()
-        if isinstance(file_, str):
-            with open(file_, encoding="utf-8") as f:
-                return s.load(f)
-        else:
-            return s.load(file_)
-
-    def copy(self) -> "Computation":
-        """Create a copy of a computation.
-
-        The copy is shallow. Any values in the new Computation's DAG will be the same object as this Computation's
-        DAG. As new objects will be created by any further computations, this should not be an issue.
-
-        :rtype: Computation
-        """
-        obj = Computation()
-        obj.dag = nx.DiGraph(self.dag)
-        obj._tag_map = defaultdict(set, {tag: nodes.copy() for tag, nodes in self._tag_map.items()})
-        obj._state_map = {state: nodes.copy() for state, nodes in self._state_map.items()}
-        return obj
-
-    @_notifies_subscribers(graph_changed=True)
-    def add_named_tuple_expansion(self, name: Name, namedtuple_type: type, group: str | None = None) -> None:
-        """Automatically add nodes to extract each element of a named tuple type.
-
-        It is often convenient for a calculation to return multiple values, and it is polite to do this a namedtuple
-        rather than a regular tuple, so that later users have same name to identify elements of the tuple. It can
-        also help make a computation clearer if a downstream computation depends on one element of such a tuple,
-        rather than the entire tuple. This does not affect the computation per se, but it does make the intention
-        clearer.
-
-        To avoid having to create many boiler-plate node definitions to expand namedtuples, the
-        ``add_named_tuple_expansion`` method automatically creates new nodes for each element of a tuple. The
-        convention is that an element called 'element', in a node called 'node' will be expanded into a new node
-        called 'node.element', and that this will be applied for each element.
-
-        Example::
-
-            >>> from collections import namedtuple
-            >>> Coordinate = namedtuple('Coordinate', ['x', 'y'])
-            >>> comp = Computation()
-            >>> comp.add_node('c', value=Coordinate(1, 2))
-            >>> comp.add_named_tuple_expansion('c', Coordinate)
-            >>> comp.compute_all()
-            >>> comp.value('c.x')
-            1
-            >>> comp.value('c.y')
-            2
-
-        :param name: Node to cera
-        :param namedtuple_type: Expected type of the node
-        :type namedtuple_type: namedtuple class
-        """
-
-        def make_f(field_name: str) -> Callable[[Any], Any]:
-            """Create a function to extract a field from a namedtuple."""
-
-            def get_field_value(tuple_val: Any) -> Any:
-                """Extract field value from the namedtuple."""
-                return getattr(tuple_val, field_name)
-
-            return get_field_value
-
-        for field_name in namedtuple_type._fields:  # type: ignore[attr-defined]
-            node_name = f"{name}.{field_name}"
-            self.add_node(node_name, make_f(field_name), kwds={"tuple_val": name}, group=group)
-            self.set_tag(node_name, SystemTags.EXPANSION)
-
-    @_notifies_subscribers(graph_changed=True)
-    def add_map_node(
-        self,
-        result_node: Name,
-        input_node: Name,
-        subgraph: "Computation",
-        subgraph_input_node: Name,
-        subgraph_output_node: Name,
-    ) -> None:
-        """Apply a graph to each element of iterable.
-
-        In turn, each element in the ``input_node`` of this graph will be inserted in turn into the subgraph's
-        ``subgraph_input_node``, then the subgraph's ``subgraph_output_node`` calculated. The resultant list, with
-        an element or each element in ``input_node``, will be inserted into ``result_node`` of this graph. In this
-        way ``add_map_node`` is similar to ``map`` in functional programming.
-
-        :param result_node: The node to place a list of results in **this** graph
-        :param input_node: The node to get a list input values from **this** graph
-        :param subgraph: The graph to use to perform calculation for each element
-        :param subgraph_input_node: The node in **subgraph** to insert each element in turn
-        :param subgraph_output_node: The node in **subgraph** to read the result for each element
-        """
-
-        def f(xs: Iterable[Any]) -> list[Any]:
-            """Apply subgraph computation to each element in the input."""
-            results: list[Any] = []
-            is_error = False
-            for x in xs:
-                subgraph.insert(subgraph_input_node, x)
-                subgraph.compute(subgraph_output_node)
-                if subgraph.state(subgraph_output_node) == States.UPTODATE:
-                    results.append(subgraph.value(subgraph_output_node))
-                else:
-                    is_error = True
-                    results.append(subgraph.copy())
-            if is_error:
-                msg = f"Unable to calculate {result_node}"
-                raise MapException(msg, results)
-            return results
-
-        self.add_node(result_node, f, kwds={"xs": input_node})
-
-    def prepend_path(self, path: Name | ConstantValue, prefix_path: NodeKey) -> NodeKey | ConstantValue:
-        """Prepend a prefix path to a node path."""
-        if isinstance(path, ConstantValue):
-            return path
-        nk = to_nodekey(path)
-        return prefix_path.join(nk)
-
-    @_notifies_subscribers(graph_changed=True)
-    def add_block(
-        self,
-        base_path: Name,
-        block: "Computation",
-        *,
-        keep_values: bool | None = True,
-        links: dict[str, Name] | None = None,
-        metadata: dict[str, Any] | None = None,
-    ) -> None:
-        """Add a computation block as a subgraph to this computation."""
-        base_path_nk = to_nodekey(base_path)
-        for node_name in block.nodes():
-            node_key = to_nodekey(node_name)
-            node_data = block.dag.nodes[node_key]
-            tags = node_data.get(NodeAttributes.TAG, None)
-            # strip the serialize tag from the original node: add_block explicitly
-            # sets serialize=False, meaning "don't serialize the function".
-            # Value serialization is controlled separately via keep_values below.
-            if tags is not None:
-                tags = tags - {SystemTags.SERIALIZE}
-            style = node_data.get(NodeAttributes.STYLE, None)
-            group = node_data.get(NodeAttributes.GROUP, None)
-            args_def, kwds_def = block.get_definition_args_kwds(node_key)
-            args_prepended = [self.prepend_path(arg, base_path_nk) for arg in args_def]
-            kwds_prepended = {k: self.prepend_path(v, base_path_nk) for k, v in kwds_def.items()}
-            func = node_data.get(NodeAttributes.FUNC, None)
-            executor = node_data.get(NodeAttributes.EXECUTOR, None)
-            converter = node_data.get(NodeAttributes.CONVERTER, None)
-            new_node_name = self.prepend_path(node_name, base_path_nk)
-            self.add_node(
-                new_node_name,
-                func,
-                args=args_prepended,
-                kwds=kwds_prepended,
-                converter=converter,
-                serialize=False,
-                inspect=False,
-                group=group,
-                tags=tags,
-                style=style,
-                executor=executor,
-            )
-            if keep_values and NodeAttributes.VALUE in node_data:
-                new_node_key = to_nodekey(new_node_name)
-                self._set_state_and_literal_value(
-                    new_node_key, node_data[NodeAttributes.STATE], node_data[NodeAttributes.VALUE]
-                )
-                # The node has a concrete value — mark it serializable so the
-                # value survives a JSON roundtrip even though the function is not.
-                self._set_tag_one(new_node_key, SystemTags.SERIALIZE)
-        if links is not None:
-            for target, source in links.items():
-                self.link(base_path_nk.join_parts(target), source)
-        if metadata is not None:
-            self._metadata[base_path_nk] = metadata
-        else:
-            if base_path_nk in self._metadata:
-                del self._metadata[base_path_nk]
-
-    @_notifies_subscribers(graph_changed=True)
-    def link(self, target: Name, source: Name) -> None:
-        """Create a link between two nodes in the computation graph."""
-        target_nk = to_nodekey(target)
-        source_nk = to_nodekey(source)
-        if target_nk == source_nk:
-            return
-
-        target_style = self._style_one(target_nk) if self.has_node(target_nk) else None
-        source_style = self._style_one(source_nk) if self.has_node(source_nk) else None
-        style = target_style if target_style else source_style
-
-        self.add_node(target_nk, identity_function, kwds={"x": source_nk}, style=style)
 
     def _repr_svg_(self) -> str | None:
         """Return SVG representation for Jupyter notebook display."""
@@ -2405,29 +526,3 @@ class Computation:
         obj = definition_class()
         populate_computation_from_class(comp, definition_class, obj, ignore_self=ignore_self)
         return comp
-
-    @_notifies_subscribers()
-    def inject_dependencies(self, dependencies: dict[Name, Any], *, force: bool = False) -> None:
-        """Injects dependencies into the nodes of the current computation where nodes are in a placeholder state.
-
-        (or all possible nodes when the 'force' parameter is set to True), using values
-        provided in the 'dependencies' dictionary.
-
-        Each key in the 'dependencies' dictionary corresponds to a node identifier, and the associated
-        value is the dependency object to inject. If the value is a callable, it will be added as a calc node.
-
-        :param dependencies: A dictionary where each key-value pair consists of a node identifier and
-                             its corresponding dependency object or a callable that returns the dependency object.
-        :param force: A boolean flag that, when set to True, forces the replacement of existing node values
-                      with the ones provided in 'dependencies', regardless of their current state. Defaults to False.
-        :return: None
-        """
-        for n in self.nodes():
-            if force or self.s[n] == States.PLACEHOLDER:
-                obj = dependencies.get(n)
-                if obj is None:
-                    continue
-                if callable(obj):
-                    self.add_node(n, obj)
-                else:
-                    self.add_node(n, value=obj)
